+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- webapp: Fix previews of smaller images
 - docker: Fix docker-compose setup
 
 ## [1.4.1] - 2022-03-26